refactor: moved IEntityFetcher to the Shesha.Framework, moved NHibernateEntityFetcher to Shesha.NHibernate
