+++ shesha-core/src/Shesha.GraphQL/Shesha.GraphQL.csproj
@@ -20,7 +20,6 @@
     <PackageReference Include="GraphQL.MicrosoftDI" Version="5.3.0" />
     <PackageReference Include="GraphQL.NewtonsoftJson" Version="5.3.0" />
     <PackageReference Include="GraphQL.SystemTextJson" Version="5.3.0" />
-    <PackageReference Include="NHibernate" Version="5.4.3" />
   </ItemGroup>
 
   <ItemGroup>
